Use long for Windows

+++ include/gz/transport/Discovery.hh
@@ -791,7 +791,7 @@ namespace ignition
         sockaddr_in clntAddr;
         socklen_t addrLen = sizeof(clntAddr);
 
-        ssize_t received = recvfrom(this->sockets.at(0),
+        long received = recvfrom(this->sockets.at(0),
               reinterpret_cast<raw_type *>(rcvStr),
               this->kMaxRcvStr, 0,
               reinterpret_cast<sockaddr *>(&clntAddr),
